optomize UI, clean code
