+++ lib/widgets/custom_combobox.dart
@@ -1,5 +1,5 @@
-import 'package:animated_custom_dropdown/custom_dropdown.dart';
 import 'package:flutter/material.dart';
+import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:viva_home_mobile/utils/constants.dart';
 
 class CustomComboBox<T> extends FormField<T> {
@@ -11,37 +11,43 @@ class CustomComboBox<T> extends FormField<T> {
     super.onSaved,
     super.validator,
     ValueChanged<T?>? onChanged,
+    bool enabled = true,
   }) : super(
          builder: (FormFieldState<T> field) {
-           return _ComboBoxField<T>(
+           return _CustomDropdownField<T>(
              field: field,
              items: items,
              hint: hint,
              onChanged: onChanged,
+             enabled: enabled,
            );
          },
        );
 }
 
-class _ComboBoxField<T> extends StatefulWidget {
+class _CustomDropdownField<T> extends StatefulWidget {
   final FormFieldState<T> field;
   final List<T> items;
   final String hint;
   final ValueChanged<T?>? onChanged;
+  final bool enabled;
 
-  const _ComboBoxField({
+  const _CustomDropdownField({
     required this.field,
     required this.items,
     required this.hint,
     this.onChanged,
+    this.enabled = true,
     super.key,
   });
 
   @override
-  State<_ComboBoxField<T>> createState() => _ComboBoxFieldState<T>();
+  State<_CustomDropdownField<T>> createState() => _CustomDropdownFieldState<T>();
 }
 
-class _ComboBoxFieldState<T> extends State<_ComboBoxField<T>> {
+class _CustomDropdownFieldState<T> extends State<_CustomDropdownField<T>> {
+  bool _isMenuOpen = false;
+
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(
@@ -49,80 +55,22 @@ class _ComboBoxFieldState<T> extends State<_ComboBoxField<T>> {
         return Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            CustomDropdown<T>.search(
-              items: widget.items,
-              onChanged: (T? value) {
-                widget.field.didChange(value);
-                widget.onChanged?.call(value);
-              },
-              hintText: "Please select an option",
-              initialItem: widget.field.value,
-              excludeSelected: false,
-              decoration: CustomDropdownDecoration(
-                closedFillColor: AppColors.white,
-                expandedFillColor: AppColors.white,
-                closedBorder: Border.all(
-                  color: widget.field.hasError
-                      ? AppColors.error
-                      : AppColors.dark,
-                  width: 0.5,
-                ),
-                expandedBorder: Border.all(
-                  color: widget.field.hasError
-                      ? AppColors.error
-                      : AppColors.primaryTeal,
-                  width: 1.5,
-                ),
-                closedBorderRadius: BorderRadius.all(
-                  Radius.circular(AppSizes.radius(context, SizeCategory.small)),
-                ),
-                expandedBorderRadius: BorderRadius.all(
-                  Radius.circular(AppSizes.radius(context, SizeCategory.small)),
-                ),
-                listItemDecoration: ListItemDecoration(
-                  selectedColor: AppColors.lightGray,
-                  splashColor: AppColors.lightGray,
-                  highlightColor: AppColors.lightGray,
-                ),
-                closedSuffixIcon: Icon(
-                  Icons.keyboard_arrow_down,
-                  color: AppColors.primaryTeal,
-                ),
-                expandedSuffixIcon: Icon(
-                  Icons.keyboard_arrow_up,
-                  color: AppColors.primaryTeal,
-                ),
+            ConstrainedBox(
+              constraints: BoxConstraints(
+                minWidth: constraints.maxWidth,
+                maxWidth: constraints.maxWidth
               ),
-              listItemBuilder: (context, item, isSelected, onItemSelect) {
-                final isLast = item == widget.items.last;
-                final isCurrentSelected = widget.field.value == item;
-
-                return Container(
-                  decoration: BoxDecoration(
-                    color: isSelected
-                        ? AppColors.lightGray
-                        : Colors.transparent,
-                    border: isLast
-                        ? null
-                        : Border(
-                            bottom: BorderSide(
-                              color: AppColors.dark,
-                              width: 0.5,
-                            ),
-                          ),
-                  ),
-                  child: InkWell(
-                    onTap: onItemSelect,
-                    child: Padding(
+              child: DropdownButtonFormField2<T>(
+                value: widget.field.value,
+                items: widget.items.map((T item) {
+                  final isSelected = widget.field.value == item;
+                  return DropdownMenuItem<T>(
+                    value: item,
+                    child: Container(
+                      width: double.infinity,
                       padding: EdgeInsets.symmetric(
-                        horizontal: AppSizes.padding(
-                          context,
-                          SizeCategory.small,
-                        ),
-                        vertical: AppSizes.padding(
-                          context,
-                          SizeCategory.small,
-                        ),
+                        horizontal: AppSizes.padding(context, SizeCategory.small) * 0.75,
+                        vertical: AppSizes.padding(context, SizeCategory.small),
                       ),
                       child: Row(
                         children: [
@@ -130,35 +78,118 @@ class _ComboBoxFieldState<T> extends State<_ComboBoxField<T>> {
                             child: Text(
                               item.toString(),
                               style: TextStyle(
-                                fontSize: AppSizes.font(
-                                  context,
-                                  SizeCategory.large,
-                                ),
-                                fontWeight: isCurrentSelected
-                                    ? FontWeight.w600
-                                    : FontWeight.w400,
-                                color: isCurrentSelected
-                                    ? AppColors.primaryTeal
-                                    : AppColors.dark,
+                                fontSize: AppSizes.font(context, SizeCategory.large),
+                                fontWeight: isSelected ? FontWeight.w600 : FontWeight.w400,
+                                color: isSelected ? AppColors.primaryTeal : AppColors.dark,
                               ),
+                              overflow: TextOverflow.ellipsis,
                             ),
                           ),
-                          if (isCurrentSelected)
+                          if (isSelected)
                             Container(
                               margin: EdgeInsets.only(left: AppSizes.padding(context, SizeCategory.small)),
                               child: Icon(
                                 Icons.check_circle,
                                 color: AppColors.primaryTeal,
-                                size: 20,
+                                size: AppSizes.icon(context, SizeCategory.medium),
                               ),
                             ),
                         ],
                       ),
                     ),
+                  );
+                }).toList(),
+                onChanged: widget.enabled ? (T? value) {
+                  widget.field.didChange(value);
+                  widget.onChanged?.call(value);
+                } : null,
+                onMenuStateChange: (isOpen) {
+                  setState(() {
+                    _isMenuOpen = isOpen;
+                  });
+                },
+                buttonStyleData: ButtonStyleData(
+                  height: AppSizes.container(context, SizeCategory.small) *0.85,
+                  width: double.infinity,
+                  padding: EdgeInsets.symmetric(horizontal: AppSizes.padding(context, SizeCategory.medium)),
+                  decoration: BoxDecoration(
+                    color: AppColors.white,
+                    border: Border.all(
+                      color: widget.field.hasError
+                          ? AppColors.error
+                          : (_isMenuOpen ? AppColors.primaryTeal : AppColors.dark),
+                      width: _isMenuOpen ? 1.5 : 1,
+                    ),
+                    
+                    borderRadius: _isMenuOpen
+                        ? BorderRadius.only(
+                            topLeft: Radius.circular(
+                              AppSizes.radius(context, SizeCategory.small),
+                            ),
+                            topRight: Radius.circular(
+                              AppSizes.radius(context, SizeCategory.small),
+                            ),
+                          )
+                        : BorderRadius.circular(
+                            AppSizes.radius(context, SizeCategory.small),
+                          ),
                   ),
-                );
-              },
+                ),
+
+                // Custom icon styling
+                iconStyleData: IconStyleData(
+                  icon: Icon(
+                    Icons.keyboard_arrow_down,
+                    color: AppColors.primaryTeal,
+                  ),
+                  openMenuIcon: Icon(
+                    Icons.keyboard_arrow_up,
+                    color: AppColors.primaryTeal,
+                  ),
+                ),
+
+                dropdownStyleData: DropdownStyleData(
+                  maxHeight: AppSizes.container(context, SizeCategory.xxlarge),
+                  width: constraints.maxWidth > 0 ? constraints.maxWidth : null,
+                  decoration: BoxDecoration(
+                    color: AppColors.white,
+                    borderRadius: BorderRadius.only(
+                      bottomLeft: Radius.circular(
+                        AppSizes.radius(context, SizeCategory.small),
+                      ),
+                      bottomRight: Radius.circular(
+                        AppSizes.radius(context, SizeCategory.small),
+                      ),
+                    ),
+                    border: Border(
+                      left: widget.field.hasError ? BorderSide(color: AppColors.error, width: 1.5) : BorderSide(color: AppColors.primaryTeal, width: 1.5),
+                      right: widget.field.hasError ? BorderSide(color: AppColors.error, width: 1.5) : BorderSide(color: AppColors.primaryTeal, width: 1.5),
+                      bottom: widget.field.hasError ? BorderSide(color: AppColors.error, width: 1.5) : BorderSide(color: AppColors.primaryTeal, width: 1.5),
+                    ),
+                  ),
+                ),
+
+                // Custom menu item styling
+                menuItemStyleData: MenuItemStyleData(
+                  height: AppSizes.container(context, SizeCategory.small) * 0.8,
+                  padding: EdgeInsets.symmetric(horizontal: AppSizes.padding(context, SizeCategory.small) * 0.75),
+                ),
+
+                // Custom validator style
+                decoration: InputDecoration(
+                  contentPadding: EdgeInsets.zero,
+                  border: InputBorder.none,
+                  enabledBorder: InputBorder.none,
+                  focusedBorder: InputBorder.none,
+                  errorBorder: InputBorder.none,
+                  focusedErrorBorder: InputBorder.none,
+                  errorStyle: TextStyle(height: 0),
+                ),
+                isExpanded: true,
+              ),
             ),
+
+            // Custom error message
             if (widget.field.hasError)
               Padding(
                 padding: EdgeInsets.only(
@@ -170,6 +201,7 @@ class _ComboBoxFieldState<T> extends State<_ComboBoxField<T>> {
                     color: AppColors.error,
                     fontSize: AppSizes.font(context, SizeCategory.medium),
                   ),
+                  overflow: TextOverflow.ellipsis,
                 ),
               ),
           ],
