+++ lib/widgets/form_widget.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:viva_home_mobile/utils/constants.dart';
+import 'package:viva_home_mobile/utils/custom_text_field.dart';
+import 'package:viva_home_mobile/utils/radio_group.dart';
 import 'package:viva_home_mobile/widgets/checkbox_individual_widget.dart';
+import 'package:viva_home_mobile/widgets/custom_combobox.dart';
 
 class FormSection extends StatelessWidget {
   final List<Widget> children;
@@ -40,7 +43,16 @@ class FormSection extends StatelessWidget {
           ),
         ],
       ),
-      child: SingleChildScrollView(child: Column(children: children)),
+      child: Padding(
+        padding: EdgeInsets.only(
+        
+          bottom: MediaQuery.of(context).viewInsets.bottom,
+        ),
+        child: SingleChildScrollView(
+          physics: const BouncingScrollPhysics(),
+          child: Column(children: children),
+        ),
+      ),
     );
   }
 }
@@ -73,7 +85,10 @@ class FormFieldWrapper extends StatelessWidget {
             mainAxisAlignment: MainAxisAlignment.start,
             children: [
               if (nodeKey != null)
-                IndividualCheckboxWidget.standalone(nodeKey: nodeKey!, showtitle: false),
+                IndividualCheckboxWidget.standalone(
+                  nodeKey: nodeKey!,
+                  showtitle: false,
+                ),
               SizedBox(width: AppSizes.padding(context, SizeCategory.small)),
               Expanded(
                 child: Text(
@@ -158,4 +173,158 @@ class FormFieldGroup extends StatelessWidget {
       ],
     );
   }
+}
+
+// Helper for ComboBox field
+FormFieldGroup buildComboBoxField({
+  onSaved,
+  validator,
+  required String label,
+  required List<String> items,
+  required Function(String?) onChanged,
+}) {
+  return FormFieldGroup(
+    label: label,
+    items: [
+      CustomComboBox<String>(
+        items: items,
+        onChanged: onChanged,
+        onSaved: onSaved,
+        validator: validator,
+      ),
+    ],
+  );
+}
+
+// Helper for TextField field
+FormFieldGroup buildCustomTextField({
+  required String label,
+  required String hintText,
+  Color colorBorder = AppColors.darkGray,
+  bool enabled = true,
+  TextInputType? keyboardType,
+  Widget? suffixIcon,
+  int? maxLength,
+  int? maxLines,
+  String? Function(String?)? validator,
+  void Function(String?)? onSaved,
+  void Function(String?)? onChanged,
+}) {
+  return FormFieldGroup(
+    label: label,
+    items: [
+      CustomTextField(
+        hintText: hintText,
+        hintTextEnable: false,
+        colorBorder: colorBorder,
+        enabled: enabled,
+        keyboardType: keyboardType ?? TextInputType.text,
+        suffixIcon: suffixIcon,
+        maxLength: maxLength,
+        maxLines: maxLines,
+        validator: validator,
+        onSaved: onSaved,
+        onChanged: onChanged,
+      ),
+    ],
+  );
+}
+
+// Helper for Radio Button field
+FormFieldGroup buildRadioField<T>({
+  required BuildContext context,
+  String? label,
+  required List<RadioOption<T>> options,
+  bool customLabelText = false,
+  required T? groupValue,
+  required Function(T?) onChanged,
+  FormFieldValidator<T>? validator,
+  FormFieldSetter<T>? onSaved,
+}) {
+  return FormFieldGroup(
+    customLabelText: customLabelText,
+    label: label ?? '',
+    items: [
+      FormField<T>(
+        initialValue: groupValue,
+        validator: validator,
+        onSaved: onSaved,
+        builder: (field) {
+          return Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              CustomRadioGroup<T>(
+                options: options,
+                groupValue: field.value,
+                onChanged: (val) {
+                  onChanged(val);
+                  field.didChange(val);
+                },
+              ),
+              if (field.hasError)
+                Padding(
+                  padding: EdgeInsets.only(
+                    top: AppSizes.container(context, SizeCategory.small) * 0.25,
+                  ),
+                  child: Text(
+                    field.errorText!,
+                    style: TextStyle(
+                      color: AppColors.error,
+                      fontSize: AppSizes.font(context, SizeCategory.medium),
+                    ),
+                  ),
+                ),
+            ],
+          );
+        },
+      ),
+    ],
+  );
+}
+
+// Helper for "If Other, please specify" field
+Widget buildOtherSpecifyField({
+  required BuildContext context,
+  String? Function(String?)? validator,
+  void Function(String?)? onSaved,
+  String highlightText = "Other",
+  String description = "please specify:",
+  String hintText = "...",
+  Color colorBorder = AppColors.dark,
+  TextInputType keyboardType = TextInputType.text,
+  bool textFieldEnabled = true,
+  int? maxLines,
+}) {
+  return FormFieldGroup(
+    customLabelText: true,
+    items: [
+      RichText(
+        text: TextSpan(
+          style: TextStyle(
+            fontSize: AppSizes.font(context, SizeCategory.large),
+            fontWeight: FontWeight.w600,
+            color: AppColors.darkTeal,
+          ),
+          children: [
+            const TextSpan(text: "If '"),
+            TextSpan(
+              text: highlightText,
+              style: TextStyle(color: AppColors.primaryTeal),
+            ),
+            TextSpan(text: "', $description"),
+          ],
+        ),
+      ),
+      if (textFieldEnabled)
+        CustomTextField(
+          validator: validator,
+          onSaved: onSaved,
+          hintText: hintText,
+          hintTextEnable: false,
+          colorBorder: colorBorder,
+          maxLines: maxLines,
+          keyboardType: keyboardType,
+        ),
+    ],
+  );
 }
